Add result, restart page countdown, button shortcuts
Add result(this will also remove the possibility to choose rock, paper
or scissors with either buttons or shortcuts), the result for winning is
a gif, while the others are plain colored text. Add a 10 second
countdown to restart the page, this countdown can be stopped with a
button that swaps images(play/pause). Add shortcuts to the rock, paper,
scissors and reset game buttons.